requirements.txt
python-dotenv
SQLAlchemy
typer
tabulate
requests

ruff